Crate scuttlebutt

Scuttlebutt

Scuttlebutt is an interface for extending kubernetes by feeding off a stream of kubernetes cluster events

Structs

Cluster
A cluster contains an address for interacting with a kubernetes Cluster of nodes

Event
Kubernets cluster event

EventSource
EventSource contains information for an event.

Object
A description of the event

ObjectMeta
ObjectMeta is metadata that all persisted resources must have, which includes all objects users must create.

ObjectReference
ObjectReference contains enough information to let you inspect or modify the referred object.

Enums

Error
An enumeratation of potential errors

Traits

Events
Events provides a means for generating a receiver for events

Type Definitions

Result