Module scuttlebutt::channel

Module for encapsulating communication channels for swanky.

Structs

Channel
A standard read/write channel that implements AbstractChannel.

HashChannel
An instantiation of the AbstractChannel trait which computes a running hash of all bytes read from and written to the channel.

SyncChannel
A channel that implements AbstractChannel as well as Send and Sync.

TrackChannel
A channel for tracking the number of bits read/written.

Traits

AbstractChannel
A trait for managing I/O. AbstractChannels are clonable, and provide basic read/write capabilities for both common and scuttlebutt-specific types.

Functions

track_unix_channel_pair
Convenience function to create a pair of TrackUnixChannels for local tests in swanky.

unix_channel_pair
Convenience function to create a pair of UnixChannels for local tests in swanky.

Type Definitions

TrackUnixChannel
A TrackChannel which uses UnixStreams.

UnixChannel
A SyncChannel which uses UnixStreams.