Trait scuttlebutt::channel::AbstractChannel

pub trait AbstractChannel: Read + Write {
    fn read_vec(&mut self, nbytes: usize) -> Result<Vec<u8>> { ... }

    fn write_bool(&mut self, b: bool) -> Result<()> { ... }

    fn read_bool(&mut self) -> Result<bool> { ... }

    fn write_u8(&mut self, s: u8) -> Result<()> { ... }

    fn read_u8(&mut self) -> Result<u8> { ... }

    fn write_u16(&mut self, s: u16) -> Result<()> { ... }

    fn read_u16(&mut self) -> Result<u16> { ... }

    fn write_u32(&mut self, s: u32) -> Result<()> { ... }

    fn read_u32(&mut self) -> Result<u32> { ... }

    fn write_u64(&mut self, s: u64) -> Result<()> { ... }

    fn read_u64(&mut self) -> Result<u64> { ... }

    fn write_usize(&mut self, s: usize) -> Result<()> { ... }

    fn read_usize(&mut self) -> Result<usize> { ... }

    fn write_block(&mut self, b: &Block) -> Result<()> { ... }

    fn read_block(&mut self) -> Result<Block> { ... }

    fn read_blocks(&mut self, n: usize) -> Result<Vec<Block>> { ... }

    fn write_block512(&mut self, b: &Block512) -> Result<()> { ... }

    fn read_block512(&mut self) -> Result<Block512> { ... }

    fn write_pt(&mut self, pt: &RistrettoPoint) -> Result<()> { ... }

    fn read_pt(&mut self) -> Result<RistrettoPoint> { ... }
}

A trait for managing I/O. AbstractChannels are essentially just Read+Write, but provide convience methods for sending and receiving both common and scuttlebutt-specific types.

Provided methods

fn read_vec(&mut self, nbytes: usize) -> Result<Vec<u8>>

Read nbytes from the channel, and return it as a Vec.

fn write_bool(&mut self, b: bool) -> Result<()>

Write a bool to the channel.

fn read_bool(&mut self) -> Result<bool>

Read a bool from the channel.

fn write_u8(&mut self, s: u8) -> Result<()>

Write a u8 to the channel.

fn read_u8(&mut self) -> Result<u8>

Read a u8 from the channel.

fn write_u16(&mut self, s: u16) -> Result<()>

Write a u16 to the channel.

fn read_u16(&mut self) -> Result<u16>

Read a u16 from the channel.

fn write_u32(&mut self, s: u32) -> Result<()>

Write a u32 to the channel.

fn read_u32(&mut self) -> Result<u32>

Read a u32 from the channel.

fn write_u64(&mut self, s: u64) -> Result<()>

Write a u64 to the channel.

fn read_u64(&mut self) -> Result<u64>

Read a u64 from the channel.

fn write_usize(&mut self, s: usize) -> Result<()>

Write a usize to the channel.

fn read_usize(&mut self) -> Result<usize>

Read a usize from the channel.

fn write_block(&mut self, b: &Block) -> Result<()>

Write a Block to the channel.

fn read_block(&mut self) -> Result<Block>

Read a Block from the channel.

fn read_blocks(&mut self, n: usize) -> Result<Vec<Block>>

Read n Blocks from the channel.

fn write_block512(&mut self, b: &Block512) -> Result<()>

Write a Block512 to the channel.

fn read_block512(&mut self) -> Result<Block512>

Read a Block512 from the channel.

fn write_pt(&mut self, pt: &RistrettoPoint) -> Result<()>

Write a RistrettoPoint to the channel.

fn read_pt(&mut self) -> Result<RistrettoPoint>

Read a RistrettoPoint from the channel.

Loading content...