Module scuttlebutt::cointoss

Implementation of a simple two-party coin tossing protocol using a PRG as a commitment.

On input seed, the sender computes r := PRG(seed) and sends r to the receiver. It then receives seed_ from the receiver and outputs seed ⊕ seed_. Likewise, on input seed, the receiver gets r, sends seed to the sender, and then receives seed_, checking that PRG(seed_) = r.

Enums

Error
Errors produced by the coin tossing protocol.

Functions

receive
Coin tossing receiver.

send
Coin tossing sender.