Module scuttlebutt::field

This module defines finite fields.

Modules

fft

Various number theoretic utility functions used in the library.

polynomial

This module defines polynomials (and their operations) over finite fields.

Structs

F2

A field element in the prime-order finite field $\textsf{GF}(2).$

F2e19x3e26

The finite field over the prime $M = 2^{19} 3^{26} + 1$. Hence, $\phi(M)$ is divisible by $2^{19}$ and $3^{26}$ and thus supports a large number of FFT<2> and FFT<3> sizes for threshold secret sharing.

F40b

An element of the finite field $\textsf{GF}(2^{40})$ reduced over $x^{40} + x^5 + x^4 + x^3 + 1$

F45b

An element of the finite field $\textsf{GF}(2^{45})$ reduced over $x^{45} + x^{28} + x^{17} + x^{11} + 1$

F56b

An element of the finite field $\textsf{GF}(2^{56})$ reduced over $x^{56} + x^8 + x^3 + x^2 + 1$

F61p

A finite field over the Mersenne Prime 2^61 - 1

F63b

An element of the finite field $\textsf{GF}(2^{63})$ reduced over $x^{63} + x + 1$

F64b

An element of the finite field $\textsf{GF}({2^{64}})$ reduced over $x^{64} + x^{19} + x^{16} + x + 1$.

F128b

An element of the finite field $\textsf{GF}(2^{128})$ reduced over $x^{128} + x^7 + x^2 + x + 1$

F128p

The finite field over the prime $P = 2^{128} - 159$.

F256p

The finite field over the prime $P = 2^{256} - 2^{224} + 2^{192} + 2^{96} - 1 = 115792089210356248762697446949407573530086143415290314195533631308867097853951$.

F384p

The finite field over the prime $P = 2^{384} - 2^{128} - 2^{96} + 2^{32} - 1 = 39402006196394479212279040100143613805079739270465446667948293404245721771496870329047266088258938001861606973112319$.

F384q

The finite field over the prime $Q = 39402006196394479212279040100143613805079739270465446667946905279627659399113263569398956308152294913554433653942643$.

F400p

The finite field over the prime $2^{400} - 593$.

Fbls12381

The BLS12-381 finite field.

Fbn254

The BN-254 finite field.

Traits

FiniteField

Types that implement this trait are finite fields.

IsSubFieldOf

Denotes that Self is a subfield of FE.

PrimeFiniteField

A PrimeFiniteField is a FiniteField with a prime modulus. In this case the field is isomorphic to integers modulo prime p.

SmallBinaryField

A GF(2) extension field such that:

Type Definitions

Degree

The degree, $r$ of a finite field.

DegreeModulo

The relative degree between two Finite Fields.