Struct scuttlebutt::AesRng

pub struct AesRng(_);

Implementation of a random number generator based on fixed-key AES.

This uses AES in a counter-mode-esque way, but with the counter always starting at zero. When used as a PRNG this is okay [TODO: citation?].

Methods

impl AesRng

pub fn new() -> Self

Create a new random number generator using a random seed from rand::random.

pub fn fork(&mut self) -> Self

Create a new RNG using a random seed from this one.