Struct scuttlebutt::Block512

pub struct Block512(_);

A 512-bit value.

Methods

impl Block512

pub fn prefix(&self, n: usize) -> &[u8]

Return the first n bytes, where n must be <= 64.

pub fn prefix_mut(&mut self, n: usize) -> &mut [u8]

Return the first n bytes as mutable, where n must be <= 64.