Struct scuttlebutt::Object

pub struct Object {
    pub api_version: String,
    pub count: usize,
    pub first_timestamp: String,
    pub last_timestamp: String,
    pub involved_object: ObjectReference,
    pub kind: String,
    pub message: String,
    pub metadata: ObjectMeta,
    pub reason: String,
    pub source: EventSource,
    pub object_type: String,
}

A description of the event

Fields
api_version: String

APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: http://releases.k8s.io/release-1.3/docs/devel/api-conventions.md#resources

count: usize

The number of times this event has occurred.

first_timestamp: String

The time at which the event was first recorded. (Time of server receipt is in TypeMeta.)

last_timestamp: String

The time at which the most recent occurrence of this event was recorded.

involved_object: ObjectReference

The object that this event is about.

kind: String

Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: http://releases.k8s.io/release-1.3/docs/devel/api-conventions.md#types-kinds

message: String

A human-readable description of the status of this operation.

metadata: ObjectMeta

Standard object’s metadata. More info: http://releases.k8s.io/release-1.3/docs/devel/api-conventions.md#metadata

reason: String

This should be a short, machine understandable string that gives the reason for the transition into the object’s current status.

source: EventSource

The component reporting this event. Should be a short machine understandable string.

object_type: String

Type of this event (Normal, Warning), new types could be added in the future