Struct scuttlebutt::ObjectReference

pub struct ObjectReference {
    pub api_version: String,
    pub resource_version: String,
    pub uid: String,
    pub field_path: Option<String>,
    pub kind: String,
    pub name: String,
    pub namespace: String,
}

ObjectReference contains enough information to let you inspect or modify the referred object.

Fields
api_version: String

API version of the referent.

resource_version: String

Specific resourceVersion to which this reference is made, if any. More info: http://releases.k8s.io/release-1.3/docs/devel/api-conventions.md#concurrency-control-and-consistency

uid: String

UID of the referent. More info: http://releases.k8s.io/release-1.3/docs/user-guide/identifiers.md#uids

field_path: Option<String>

If referring to a piece of an object instead of an entire object, this string should contain a valid JSON/Go field access statement, such as desiredState.manifest.containers[2]. For example, if the object reference is to a container within a pod, this would take on a value like: "spec.containers{name}" (where "name" refers to the name of the container that triggered the event) or if no container name is specified "spec.containers[2]" (container with index 2 in this pod). This syntax is chosen only to have some well-defined way of referencing a part of an object.

kind: String

Kind of the referent. More info: http://releases.k8s.io/release-1.3/docs/devel/api-conventions.md#types-kinds

name: String

Name of the referent. More info: http://releases.k8s.io/release-1.3/docs/user-guide/identifiers.md#names

namespace: String

Namespace of the referent. More info: http://releases.k8s.io/release-1.3/docs/user-guide/namespaces.md