Struct scuttlebutt::TrackChannel

pub struct TrackChannel<R: Read, W: Write> { /* fields omitted */ }

A channel for tracking the number of bits read/written.

Methods

impl<R: Read, W: Write> TrackChannel<R, W>

pub fn new(reader: R, writer: W) -> Self

Make a new TrackChannel from a reader and a writer.

pub fn clear(&self)

Clear the number of bits read/written.

pub fn kilobits_written(&self) -> f64

Return the number of kilobits written to the channel.

pub fn kilobits_read(&self) -> f64

Return the number of kilobits read from the channel.