Module scuttlebutt::utils

Useful utility functions.

Functions

and
AND two byte arrays, outputting the result.

and_inplace
AND two byte arrays in place.

pack_bits
Pack a bit slice into bytes.

unpack_bits
Unpack a bit vector from a slice of bytes.

xor
XOR two byte arrays, outputting the result.

xor_inplace
XOR two byte arrays in place.

xor_inplace_n
XOR two byte arrays up to n bytes in place.

xor_n
XOR two byte arrays up to n bytes, outputting the result.