Crate scuttlebutt

Scuttlebutt provides many utility functions for cryptographic applications.

Re-exports

pub use crate::channel::AbstractChannel;

pub use crate::channel::Channel;

pub use crate::channel::SymChannel;

Modules

bloomfilter

Implementation of a bloom filter.

channel

Module for encapsulating communication channels for swanky.

cointoss

Implementation of a simple two-party coin tossing protocol using a PRG as a commitment.

commitment

A trait defining a Commitment Scheme and an implementation in the random oracle model using SHA256.

field

This module defines finite fields.

generic_array_length

Helper types to make using the generic-array easier.

ring

This module defines finite rings.

serialization

Serialization types for algebraic structures.

utils

Useful utility functions.

Macros

prime_field_using_ff

This macro constructs a prime finite field using the ff library. The modulus and generator should be listed, along with the name, in build.rs.

Structs

Aes128

AES-128, encryption only.

Aes256

AES-256, encryption only.

AesHash

AES-based correlation-robust hash function.

AesRng

Implementation of a random number generator based on fixed-key AES.

Block

A 128-bit chunk.

Block512

A 512-bit value.

HashChannel

An instantiation of the AbstractChannel trait which computes a running hash of all bytes read from and written to the channel.

SyncChannel

A channel that implements AbstractChannel as well as Send and Sync.

TrackChannel

A channel wrapping another channel for tracking the number of bits read/written.

UniformIntegersUnderBound

Sample u32s uniformly from [0, bound).

Constants

AES_HASH

AesHash with a fixed key.

FIXED_KEY_AES128

Fixed-key AES-128.

Traits

Malicious

A marker trait denoting that the given scheme is maliciously secure.

SemiHonest

A marker trait denoting that the given scheme is semi-honest secure.

Functions

track_unix_channel_pair

Convenience function to create a pair of TrackUnixChannels for local tests in swanky.

unix_channel_pair

Convenience function to create a pair of UnixChannels for local tests in swanky.

Type Definitions

TrackUnixChannel

A TrackChannel which uses UnixStreams.

UnixChannel

A SyncChannel which uses UnixStreams.